Add files via upload
Hi! I have added two files. The Jupyter Notebook is a detailed explanation of the methodology used; the Julia file is just the code.
I do *not* recommend to run the codes if you need your computer: it takes a lot of time, I usually run it at night.
Cheers!